Move notification handling to helpers.js
Move it out of the way
