+++ lib/default_dialog_queue.js
@@ -12,7 +12,6 @@
 const assert = require('assert');
 
 const permissionGrant = require('./dialogs/permission_grant');
-const { showNotification, showError } = require('./dialogs/notifications');
 
 const { Intent, DispatchResult } = require('./semantic');
 const Helpers = require('./helpers');
@@ -136,7 +135,7 @@ QueueItem.Notification = class Notification extends QueueItem {
     async _doDispatch(dlg) {
         const previousAppId = dlg.currentAppId;
         dlg.currentAppId = this.appId;
-        return showNotification(dlg, this.appId, this.icon, this.outputType, this.outputValue, previousAppId);
+        return Helpers.showNotification(dlg, this.appId, this.icon, this.outputType, this.outputValue, previousAppId);
     }
 };
 
@@ -150,7 +149,7 @@ QueueItem.Error = class Error extends QueueItem {
 
     async _doDispatch(dlg) {
         dlg.currentAppId = this.appId;
-        return showError(dlg, this.appId, this.icon, this.error);
+        return Helpers.showError(dlg, this.appId, this.icon, this.error);
     }
 };
