+++ lib/helpers.js
@@ -27,7 +27,133 @@ function capitalize(str) {
     return (str[0].toUpperCase() + str.substr(1)).replace(/[.\-_]([a-z])/g, (whole, char) => ' ' + char.toUpperCase()).replace(/[.\-_]/g, '');
 }
 
+function presentExample(dlg, utterance) {
+    // on Android, we have app-level slot filling which is more powerful, so we don't
+    // want to lose the argument name information
+    if (dlg.manager.platform.type === 'android' || dlg.manager.platform.type === 'test')
+        utterance = utterance.split(' ').map((t) => t.startsWith('$') ? normalizeSlot(t) : t).join(' ');
+    else
+        utterance = utterance.split(' ').map((t) => t.startsWith('$') ? '____' : t).join(' ');
+    if (utterance.startsWith(', '))
+        utterance = utterance.substring(2);
+    return utterance;
+}
+
+function loadOneExample(dlg, ex) {
+    // refuse to slot fill pictures
+    for (let name in ex.args) {
+        let type = ex.args[name];
+        // avoid examples such as "post __" for both text and picture (should be "post picture" without slot for picture)
+        if (type.isEntity && type.type === 'tt:picture')
+            return null;
+    }
+
+    // turn the declaration into a program
+    let newprogram = ex.toProgram();
+    let slots = [];
+    let slotTypes = {};
+    for (let name in ex.args) {
+        slotTypes[name] = String(ex.args[name]);
+        slots.push(name);
+    }
+
+    let code = ThingTalk.NNSyntax.toNN(newprogram, {});
+    let monitorable;
+    if (ex.type === 'stream')
+        monitorable = true;
+    else if (ex.type === 'action')
+        monitorable = false;
+    else if (ex.type === 'query')
+        monitorable = ex.value.schema.is_monitorable;
+    else
+        monitorable = false;
+    return { utterance: ex.utterances[0],
+             type: ex.type,
+             monitorable: monitorable,
+             target: {
+                example_id: ex.id, code: code, entities: {}, slotTypes: slotTypes, slots: slots } };
+}
+
+async function showNotification(dlg, appId, icon, outputType, outputValue, previousAppId) {
+    let app;
+    if (appId !== undefined)
+        app = dlg.manager.apps.getApp(appId);
+    else
+        app = undefined;
+
+    let messages;
+    if (outputType !== null)
+        messages = await dlg.formatter.formatForType(outputType, outputValue, 'messages');
+    else
+        messages = outputValue;
+    if (!Array.isArray(messages))
+        messages = [messages];
+
+    let notifyOne = async (message) => {
+        if (typeof message === 'string')
+            message = { type: 'text', text: message };
+
+        if (typeof message !== 'object')
+            return;
+
+        if (message.type === 'text') {
+            await dlg.reply(message.text, icon);
+        } else if (message.type === 'picture') {
+            if (message.url === undefined)
+                await dlg.reply("Sorry, I can't find the picture you want.", icon);
+            else
+                await dlg.replyPicture(message.url, icon);
+        } else if (message.type === 'rdl') {
+            await dlg.replyRDL(message, icon);
+        } else if (message.type === 'button') {
+            await dlg.replyButton(message.text, message.json);
+        } else if (message.type === 'program') {
+            const loaded = loadOneExample(dlg, message.program);
+            await dlg.replyButton(presentExample(dlg, loaded.utterance), loaded.target);
+        } else {
+            await dlg.replyResult(message, icon);
+        }
+    };
+    if (app !== undefined && app.isRunning && appId !== previousAppId &&
+        (messages.length === 1 && (typeof messages[0] === 'string' || messages[0].type === 'text'))) {
+        const msg = typeof messages[0] === 'string' ? messages[0] : messages[0].text;
+        await dlg.reply(dlg._("Notification from %s: %s").format(app.name, msg), icon);
+    } else {
+        if (app !== undefined && app.isRunning && appId !== previousAppId)
+            await dlg.reply(dlg._("Notification from %s").format(app.name), icon);
+        for (let msg of messages)
+            await notifyOne(msg);
+    }
+}
+
+async function showError(dlg, appId, icon, error) {
+    let app;
+    if (appId !== undefined)
+        app = dlg.manager.apps.getApp(appId);
+    else
+        app = undefined;
+
+    let errorMessage;
+    if (typeof error === 'string')
+        errorMessage = error;
+    else if (error.name === 'SyntaxError')
+        errorMessage = dlg._("Syntax error at %s line %d: %s").format(error.fileName, error.lineNumber, error.message);
+    else if (error.message)
+        errorMessage = error.message;
+    else
+        errorMessage = String(error);
+    console.log('Error from ' + appId, error);
+
+    if (app !== undefined && app.isRunning)
+        await dlg.reply(dlg._("%s had an error: %s.").format(app.name, errorMessage), icon);
+    else
+        await dlg.reply(dlg._("Sorry, that did not work: %s.").format(errorMessage), icon);
+}
+
 module.exports = {
+    showNotification,
+    showError,
+
     formatError(e) {
         if (e.name === 'SyntaxError')
             return "Syntax error at line " + e.lineNumber + ": " + e.message;
@@ -118,52 +244,8 @@ module.exports = {
         });
     },
 
-    presentExample(dlg, utterance) {
-        // on Android, we have app-level slot filling which is more powerful, so we don't
-        // want to lose the argument name information
-        if (dlg.manager.platform.type === 'android' || dlg.manager.platform.type === 'test')
-            utterance = utterance.split(' ').map((t) => t.startsWith('$') ? normalizeSlot(t) : t).join(' ');
-        else
-            utterance = utterance.split(' ').map((t) => t.startsWith('$') ? '____' : t).join(' ');
-        if (utterance.startsWith(', '))
-            utterance = utterance.substring(2);
-        return utterance;
-    },
-
-    loadOneExample(dlg, ex) {
-        // refuse to slot fill pictures
-        for (let name in ex.args) {
-            let type = ex.args[name];
-            // avoid examples such as "post __" for both text and picture (should be "post picture" without slot for picture)
-            if (type.isEntity && type.type === 'tt:picture')
-                return null;
-        }
-
-        // turn the declaration into a program
-        let newprogram = ex.toProgram();
-        let slots = [];
-        let slotTypes = {};
-        for (let name in ex.args) {
-            slotTypes[name] = String(ex.args[name]);
-            slots.push(name);
-        }
-
-        let code = ThingTalk.NNSyntax.toNN(newprogram, {});
-        let monitorable;
-        if (ex.type === 'stream')
-            monitorable = true;
-        else if (ex.type === 'action')
-            monitorable = false;
-        else if (ex.type === 'query')
-            monitorable = ex.value.schema.is_monitorable;
-        else
-            monitorable = false;
-        return { utterance: ex.utterances[0],
-                 type: ex.type,
-                 monitorable: monitorable,
-                 target: {
-                    example_id: ex.id, code: code, entities: {}, slotTypes: slotTypes, slots: slots } };
-    },
+    presentExample,
+    loadOneExample,
 
     async loadExamples(dlg, dataset, maxCount) {
         const parsed = await ThingTalk.Grammar.parseAndTypecheck(dataset, dlg.manager.schemas);
