+++ lib/stack-frames/result.js
@@ -12,7 +12,7 @@
 const { Intent, ValueCategory } = require('../semantic');
 const DialogStackFrame = require('./base');
 
-const { showNotification, showError } = require('../dialogs/notifications');
+const Helpers = require('../helpers');
 const AsyncQuestionStackFrame = require('./async-question');
 
 const PAGE_SIZE = 5;
@@ -85,9 +85,9 @@ module.exports = class ResultStackFrame extends DialogStackFrame {
             const next = this._results[i];
 
             if (next.isNotification)
-                await showNotification(this._dlg, undefined, next.icon, next.outputType, next.outputValue, next.currentChannel, undefined);
+                await Helpers.showNotification(this._dlg, undefined, next.icon, next.outputType, next.outputValue, next.currentChannel, undefined);
             else if (next.isError)
-                await showError(this._dlg, undefined, next.icon, next.error, undefined);
+                await Helpers.showError(this._dlg, undefined, next.icon, next.error, undefined);
         }
         this._index = Math.min(this._index + PAGE_SIZE, this._results.length);
